Conflicts:
	YGEGame/media/level/level01.lev
	YGEGame/media/level/level02.lev
	YGEGame/media/level/level03.lev
	YGEGame/src/YGEGame/GameBall.h
	YGEGame/src/YGEGame/GameGyrocopter.cpp
	YGEGame/src/YGEGame/GameHUD.h
	YGEGame/src/YGELib/YGEEngineCore.h
	YGEGame/src/YGELib/YGEEngineCoreSingleThreaded.cpp
	YGEGame/src/YGELib/YGEEngineCoreSingleThreaded.h
	YGEGame/src/YGELib/YGEQuaternion.h
	YGEGame/src/YGELib/YGESpace.cpp

+++ YGEGame/src/YGELib/YGETimer.cpp
@@ -4,26 +4,26 @@
 namespace YGECore {
 
 
-YGETimer::YGETimer(){
-	basetime = SDL_GetTicks() * 1000;
-}
+	YGETimer::YGETimer(){
+		basetime = SDL_GetTicks() * 1000;
+	}
 
-long long YGETimer::getTime(){
-	return SDL_GetTicks() * 1000 - basetime;
-}
+	long long YGETimer::getTime(){
+		return SDL_GetTicks() * 1000 - basetime;
+	}
 
-long YGETimer::getTimeMilliseconds(){
-	return getTime() / 1000;
-}
+	long YGETimer::getTimeMilliseconds(){
+		return long(getTime() / 1000);
+	}
 
-void YGETimer::startTimer(){
-	basetime = SDL_GetTicks() * 1000;
+	void YGETimer::startTimer(){
+		basetime = SDL_GetTicks() * 1000;
 
-}
+	}
 
-long long YGETimer::stopTimer(){
-	return stoppedTime = getTime();
-}
+	long long YGETimer::stopTimer(){
+		return stoppedTime = getTime();
+	}
